Add files via upload
In our study titled "Identification of Cancer-Causing Genes from Microarray Data Using Metaheuristics and Machine Learning," we conducted a comprehensive analysis of three gene expression datasets: acute leukemia, lymphoma, and prostate cancer. To train and validate our models, we divided each dataset into training and testing sets. The training sets, encompassing 80% of the data, were used to build the models. The remaining 20% of the data served as the testing sets to evaluate the models' performance.